+++ tranquilizer/__init__.py
@@ -1,4 +1,4 @@
 '''Tranquilizer'''
 from .decorator import tranquilize
 
-__version__ = '0.3.2'
+__version__ = '0.3.4'